NODE-189 Fixed IDEA issues with Generator
